IGNITE-7993 Added IGNITE_STRIPED_POOL_DISABLED system property
* Renamed StripedExecutor to StripedExecutorImpl
* Added StripedExecutor interface
* Update StripedExecutorProxy (extends IgniteThreadPoolExecutor impl StripedExecutor)
