+++ modules/core/src/main/java/org/apache/ignite/IgniteSystemProperties.java
@@ -831,6 +831,12 @@ public final class IgniteSystemProperties {
     public static final String IGNITE_LOADED_PAGES_BACKWARD_SHIFT_MAP = "IGNITE_LOADED_PAGES_BACKWARD_SHIFT_MAP";
 
     /**
+     * When set to {@code true}, StripedExecutorProxy (IgniteThreadPoolExecutor) will be used instread of StripedExecutorImpl.
+     * Default is {@code false}.
+     */
+    public static final String IGNITE_STRIPED_POOL_DISABLED = "IGNITE_STRIPED_POOL_DISABLED";
+
+    /**
      * Enforces singleton.
      */
     private IgniteSystemProperties() {
