+++ modules/core/src/test/java/org/apache/ignite/internal/util/StripedExecutorTest.java
@@ -29,7 +29,7 @@ public class StripedExecutorTest extends GridCommonAbstractTest {
 
     /** {@inheritDoc} */
     @Override public void beforeTest() {
-        stripedExecSvc = new StripedExecutor(3, "foo name", "pool name", new JavaLogger());
+        stripedExecSvc = new StripedExecutorImpl(3, "foo name", "pool name", new JavaLogger());
     }
 
     /** {@inheritDoc} */
